--- FlyBird-JP/client/index.js
@@ -1,3 +1,0 @@
-"use strict";
-
-import '/modules/button-login.js'
