+++ FlyBird-JP/client/modules/index.js
@@ -1,0 +1,3 @@
+"use strict";
+
+import '/modules/button-auth.js'
